Added the option force
Now, the default behavior is the -i option of rm. The user is prompted for deleting files.
You can change this by passing the -f option to clean, skipping the prompt and deleting files without any questions.

+++ clean.py
@@ -28,9 +28,19 @@ def list_files_to_clean(pattern_list, file_list):
     """Return all the files in file_list that match at least one of the patterns in pattern_list"""
     return [file for pattern in pattern_list for file in file_list if fnmatch.fnmatch(file, pattern)]
 
-def remove(to_clean):
+def ask_user_before_remove(file):
+    prompt = "Do you want to remove {} ? "
+    prompt = prompt.format(file)
+    answer = input(prompt)
+    while (answer != "n" and answer != "y"):
+        print("Please answer y for yes or n for no.")
+        answer = input(prompt)
+    return (answer == 'y')
+
+def remove(to_clean, force):
     for file in to_clean:
-        os.remove(file)
+        if (not force and ask_user_before_remove(file)):
+            os.remove(file)
 
 def run(args):
     """run the program"""
@@ -38,13 +48,14 @@ def run(args):
     cible = os.getenv("PWD", '.')
     file_list = list_file_fullpath(cible)
     to_clean = list_files_to_clean(pattern_list, file_list)
-    remove(to_clean)
+    remove(to_clean, args.force)
 
 def main():
     """Parse the arguments and pass it to run"""
     version = 0.1
     parser = argparse.ArgumentParser(prog="clean")
     parser.add_argument("--version", action="version", version="%(prog)s {}".format(version))
+    parser.add_argument("-f", "--force", action="store_true", help="Don't prompt the user before removal")
     run(parser.parse_args())
 
 if __name__ == '__main__':
